fix 修复 SysOssConfig 主键类型错误

+++ ruoyi-system/src/main/java/com/ruoyi/system/domain/SysOssConfig.java
@@ -22,7 +22,7 @@ public class SysOssConfig extends BaseEntity {
      * 主建
      */
     @TableId(value = "oss_config_id")
-    private Integer ossConfigId;
+    private Long ossConfigId;
 
     /**
      * 配置key
